Fix the z-index of the MLH badge and navigation menu
Before, the items in the FAQ and Sponsoring sections would appear
on top of the navigation menu, and the MLH badge is below those
items. This change introduces fixes such that the navigation menu
is always on top of everything else, and the MLH badge on top of
the section items.